+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # rlang 0.2.2.9000
 
+* `enexprs()` and `enquos()` now fully support the `.ignore_empty`
+  option, even with named arguments (#414).
+
 * The `call` argument of `abort()` and condition constructors is now
   deprecated in favour of storing full backtraces.